Added first approach to change validity check for forms
