+++ src/helpers/form-validation/actions.js
@@ -1,5 +1,4 @@
 export const SET_FORM_VALUE = 'SET_FORM_VALUE';
-export const SET_FORM_VALIDITY = 'SET_FORM_VALIDITY';
 
 export const setFormValue = (newFormValue) => {
   return {
@@ -7,10 +6,3 @@ export const setFormValue = (newFormValue) => {
     payload: newFormValue
    };
 };
-
-export const setFormValidity = (values) => {
-  return {
-    type: SET_FORM_VALIDITY,
-    payload: values
-  }
-}
