+++ src/helpers/form-validation/reducer.js
@@ -1,4 +1,5 @@
 import { SET_FORM_VALUE, SET_FORM_VALIDITY } from './actions';
+import { getModelValidity } from './helpers';
 
 const formReducer = (state, action) => {
   const { type, payload } = action;
@@ -6,18 +7,15 @@ const formReducer = (state, action) => {
   switch (type) {
     case SET_FORM_VALUE :
       const { name, value } = payload;
-      return {
-        ...state,
-        values: {
-          ...state.values,
-          [name]: value
-        }
-      };
-    case SET_FORM_VALIDITY :
-      return {
-        ...state,
-        isModelValid: !state.isModelValid
+      const newValues = {
+        ...state.values,
+        [name]: value
       }
+      return {
+        ...state,
+        values: newValues,
+        isModelValid: getModelValidity({ validation: state.validation, values: newValues })
+      };
     default:
       return state
   }
